ml/requirements.txt
flask>=3.0.0
gunicorn>=21.2.0
pdfplumber>=0.11.0
python-docx>=1.1.0
scikit-learn>=1.4.0
pandas>=2.2.0
numpy>=1.26.0
mysql-connector-python>=8.3.0
